Don't use external annotations in compiler
Postpone the decision of using them in the compiler or IDE until later.

This fixes LazyJavaAnnotations which was breaking the contract of Annotations:
findAnnotation(FqName) was looking for external annotations, while iterator()
did not. This resulted in some inconsistencies in compiler and IDE tests.

The other way, i.e. making iterator() look up external annotations, would
have been too costly and with no clear benefit at the moment.